Finalized code for first release.
This commit reflects a lot of fine-tuning to prepare the code for the first release. Since the last commit, the code was reorganized into directories and split up into smaller modules for readability. Icon files were generated for the Windows and Mac standalone applications. Mac OS-specific exception-handling was added to the code to make GUI widgets look nicer on Mac computers. The code as it is now was used to compile both Windows and Mac standalone applications of the program.